[DC-1019] reverting client building location change
