+++ data_steward/ci/test_datasets.py
@@ -1,8 +1,29 @@
 """
 Modules for setting up and tearing down test datasets for integration testing.
 """
-from ci.setup import get_client
+from google.oauth2 import service_account
+from google.cloud import bigquery
+
 from utils import bq
+
+CLIENT = None
+"""Bigquery  client object"""
+
+
+def get_client(project_id, app_creds):
+    """
+    Ensure only one client is created and reused
+
+    :param project_id:  project to get a client for
+    :returns: a big query client object
+    """
+    global CLIENT
+    if not CLIENT:
+        credentials = service_account.Credentials.from_service_account_file(
+            app_creds)
+        CLIENT = bigquery.Client(project=project_id, credentials=credentials)
+
+    return CLIENT
 
 
 def create_dataset(project, dataset_id, description, tags, app_creds):
